==> 1_download_countries_data.ipynb <==
import json
import requests
import copy
import pandas as pd

abc = [chr(ord('a') + c) for c in range(26)]

country_codes = []
for l in abc:
    for r in abc:
        country_codes.append(l+r)

data_countries = []
for code in country_codes:
    try:
        country = requests.get("http://api.worldbank.org/v2/country/%s?format=json" % code).json()
        if len(country) > 1:
            data_countries.append(country)
    except Exception as exc:
        print('Could not fetch or decode! [%s]' % exc)

valid_countries = []
for e in data_countries:
    x = e[1][0]
    x['region_id'] = x['region']['id']
    x['region_iso2code'] = x['region']['iso2code']
    x['region_value'] = x['region']['value']
    x['adminregion_id'] = x['adminregion']['id']
    x['adminregion_iso2code'] = x['adminregion']['iso2code']
    x['adminregion_value'] = x['adminregion']['value']
    x['incomeLevel_id'] = x['incomeLevel']['id']
    x['incomeLevel_iso2code'] = x['incomeLevel']['iso2code']
    x['incomeLevel_value'] = x['incomeLevel']['value']
    x['lendingType_id'] = x['lendingType']['id']
    x['lendingType_iso2code'] = x['lendingType']['iso2code']
    x['lendingType_value'] = x['lendingType']['value']
    x.pop('region')
    x.pop('adminregion')
    x.pop('incomeLevel')
    x.pop('lendingType')
    valid_countries.append(x)

countries_df = pd.DataFrame(valid_countries)

countries_df.to_csv('contries_data_from_world_bank.csv', index=False)
