==> 2_data_preprocessing.ipynb <==
### Plan of the notebook


1. Remove columns id, iso2Code, capitalCity, lendingType_*
2. Remove records with NA region
3. All columns region_* and adminRegion_* transform into one-hot-encoding over values of region_value
4. All columns incomeLevel_* transform into one ordinal column
5. Check for missing values
6. Save clean data frame to csv

import pandas as pd
import numpy as np

countries_df = pd.read_csv('contries_data_from_world_bank.csv')

countries_df

countries_df[countries_df['iso2Code'] == 'KP']

countries_df['incomeLevel_id'].unique()

countries_df['lendingType_value'].unique()

countries_df['adminregion_value'].unique()

countries_df['region_value'].unique()

#### 1. Remove columns id, iso2Code, capitalCity, lendingType_*

columns_to_drop = ['id', 'iso2Code', 'capitalCity', 'lendingType_id', 'lendingType_iso2code', 'lendingType_value']
countries_1 = countries_df.drop(columns_to_drop, axis = 1)

countries_1

#### 2. Remove records with NA region

countires_2 = countries_1[countries_1['region_value'] != 'Aggregates']

countires_2

#### 3. All columns region_* and adminRegion_* transform into one-hot-encoding over values of region_value

dummies = pd.get_dummies(countires_2['region_value'], prefix= None)

dummies

region_columns_to_drop = ['region_id', 'region_iso2code', 'region_value',  'adminregion_id', 
                          'adminregion_iso2code', 'adminregion_value' ]

countries_no_region_columns = countires_2.drop(region_columns_to_drop, axis = 1)

countries_clean_with_dummies = pd.merge(countries_no_region_columns, dummies, left_index=True, right_index=True)

countries_clean_with_dummies

#### 4. All columns incomeLevel_* transform into one ordinal column

def il_ordinal(row):
    if row['incomeLevel_id'] == 'LIC':
        return 1
    if row['incomeLevel_id'] == 'LMC':
        return 2
    if row['incomeLevel_id'] == 'UMC':
        return 3
    if row['incomeLevel_id'] == 'HIC':
        return 4

countries_clean_with_dummies['income_level'] = countries_clean_with_dummies.apply(lambda row: il_ordinal(row), axis=1)

countries_clean_with_dummies

income_columns_to_drop = ['incomeLevel_id', 'incomeLevel_iso2code', 'incomeLevel_value']
countries_clean = countries_clean_with_dummies.drop(income_columns_to_drop, axis = 1)

countries_clean

#### 5. Check for missing values

countries_clean.isnull().sum()

data_preprocessed = countries_clean[countries_clean['longitude'].notnull()]

data_preprocessed

#### 6. Save clean data frame to csv

data_preprocessed.to_csv('countries_data_preprocessed.csv', index = False)
