==> 3_find_knn.ipynb <==
### Plan of the notebook

1. Use kNN to find k most similar countries for North Korea

import pandas as pd
import numpy as  np

from sklearn.neighbors import NearestNeighbors
from sklearn.preprocessing import StandardScaler

k_nearest_countries = 6
country_to_estimate = 'Korea, Dem. People’s Rep.'

data_raw = pd.read_csv('countries_data_preprocessed.csv')

data_raw

conutry_to_estimate_index = data_raw[data_raw['name'] == country_to_estimate].index.asi8[0]

conutry_to_estimate_index

#### 1. Use kNN to find k most similar countries for North Korea

country_names = data_raw['name']

data = data_raw.drop(['name'], axis = 1)

data

standardizer = StandardScaler()

data_standardized = standardizer.fit_transform(data)

data_standardized

data_standardized.shape

# keep and remove the country we want to estimate for
record_of_country_to_estimate = data_standardized[conutry_to_estimate_index]
data_standardized_no_country_to_estimate = np.delete(data_standardized, conutry_to_estimate_index, 0)

record_of_country_to_estimate

data_standardized_no_country_to_estimate.shape

nearest_countries_result = NearestNeighbors(n_neighbors=k_nearest_countries).fit(data_standardized_no_country_to_estimate)

distances, indices = nearest_countries_result.kneighbors([record_of_country_to_estimate])

distances

indices

nearest_countries_list = country_names[np.array(indices[0])]

nearest_countries_list

nearest_countries_list.to_csv('nearest_countries.csv', index=False)
