==> 4_download_covid_data.ipynb <==
import pandas as pd
import json
import requests

covid_raw_data = requests.get("https://pomber.github.io/covid19/timeseries.json").json()

all_records = []
for country, records in covid_raw_data.items():
    for record in records:
        record['country'] = country
        all_records.append(record)

covid_data = pd.DataFrame(all_records)

covid_data

covid_data.to_csv('covid_data.csv', index = False)
