==> 4_5_countries_mapping.ipynb <==
import pandas as pd
import numpy as np
import json

covid_data = pd.read_csv('covid_data.csv')

covid_data

names_covid = covid_data['country'].unique()

countries_data = pd.read_csv('countries_data_preprocessed.csv')

countries_data

names_countries = countries_data['name'].unique()

diff_names_covid = np.setdiff1d(names_covid, names_countries)

diff_names_covid

diff_names_countries = np.setdiff1d(names_countries, names_covid)

diff_names_countries

common_names = np.intersect1d(names_covid, names_countries)

common_names

countries_mapping = {name:name for name in common_names}

countries_mapping['Bahamas, The'] = 'Bahamas'
countries_mapping['Brunei Darussalam'] = 'Brunei'
countries_mapping['Myanmar'] = 'Burma'
countries_mapping['Congo, Rep.'] = 'Congo (Brazzaville)'
countries_mapping['Congo, Dem. Rep.'] = 'Congo (Kinshasa)'
countries_mapping['Czech Republic'] = 'Czechia'
countries_mapping['Egypt, Arab Rep.'] = 'Egypt'
countries_mapping['Gambia, The'] = 'Gambia'
countries_mapping['Iran, Islamic Rep.'] = 'Iran'
countries_mapping['Korea, Rep.'] = 'Korea, South'
countries_mapping['Kyrgyz Republic'] = 'Kyrgyzstan'
countries_mapping['Lao PDR'] = 'Laos'
countries_mapping['Russian Federation'] = 'Russia'
countries_mapping['St. Kitts and Nevis'] = 'Saint Kitts and Nevis'
countries_mapping['St. Lucia'] = 'Saint Lucia'
countries_mapping['St. Vincent and the Grenadines'] = 'Saint Vincent and the Grenadines'
countries_mapping['Slovak Republic'] = 'Slovakia'
countries_mapping['Syrian Arab Republic'] = 'Syria'
countries_mapping['United States'] = 'US'
countries_mapping['Venezuela, RB'] = 'Venezuela'
countries_mapping['Yemen, Rep.'] = 'Yemen'

countries_mapping

fh = open('countries_mapping.json', 'w')
json.dump(countries_mapping, fh)
fh.close()
