==> 5_make_an_estimation_of_confirmed_cases.ipynb <==
import pandas as pd
import json

date_of_prediction = '2020-5-1'

covid_data = pd.read_csv('covid_data.csv')

covid_data

covid_data_on_date_of_prediction = covid_data[covid_data['date'] == date_of_prediction].drop(['date'], axis = 1)

covid_data_on_date_of_prediction

k_nearest_countries = pd.read_csv('nearest_countries.csv')

k_nearest_countries

fh = open('countries_mapping.json', 'r')
countries_mapping = json.load(fh)
fh.close()

countries_mapping

def map_country_name(row):
    if row['name'] in countries_mapping:
        return countries_mapping[row['name']]
    else:
        return ''

k_nearest_countries_mapped_names = \
    pd.DataFrame(k_nearest_countries.apply(lambda row: map_country_name(row), axis=1), columns=['country'])

k_nearest_countries_mapped_names

nearest_countries_on_date_of_prediction = pd.merge(k_nearest_countries_mapped_names, covid_data_on_date_of_prediction, on='country')

nearest_countries_on_date_of_prediction

nearest_countries_on_date_of_prediction.mean()

nearest_countries_on_date_of_prediction.median()
